+++ ui/src/components/actionsheet.vue
@@ -1,17 +1,17 @@
 <template lang="pug">
 transition(name="fade" @enter="enter")
   div(v-if="showButton")
-    .wrapper.add-new-item.d-flex.flex-wrap.justify-content-center.align-content-center.add-new-item-hover.pulse
-      b-icon.add-icon.show-icon(icon="gear-wide-connected", animation="spin", aria-hidden="true", @click="show = !show")
+    .wrapper.add-new-item.d-flex.flex-wrap.justify-content-center.align-content-center.add-new-item-hover.pulse(@click="show = !show")
+      b-icon.add-icon.show-icon(icon="gear-wide-connected", animation="spin", aria-hidden="true")
     b-modal(v-model="show" modal-class="overflow-hidden" dialog-class="modal-bottom" hide-footer hide-header)
-      .row.pb-4(v-if="showLeftNavi || showRightNavi")
-        .col.d-flex.flex-wrap.justify-content-start.align-content-center
+      .row.pb-4(v-if="showLeftNavi || showRightNavi || showMiddleNavi")
+        .col.d-flex.flex-wrap.justify-content-start.align-content-center(v-if="showLeftNavi")
           b-button.left-button(pill, @click="$emit('leftNaviClick')") {{ leftNaviName }}
-        .col.d-flex.flex-wrap.justify-content-center.align-content-center
+        .col.d-flex.flex-wrap.justify-content-center.align-content-center(v-if="showMiddleNavi")
           b-button.middle-button(pill, @click="$emit('middleNaviClick')") {{ middleNaviName }}
-        .col.d-flex.flex-wrap.justify-content-end.align-content-center
+        .col.d-flex.flex-wrap.justify-content-end.align-content-center(v-if="showRightNavi")
           b-button.btn-primary.right-button(pill, @click="$emit('rightNaviClick')") {{ rightNaviName }}
-      hr.mt-0.pt-0(v-if="showLeftNavi || showRightNavi")
+      hr.mt-0.pt-0(v-if="showLeftNavi || showRightNavi || showMiddleNavi")
       div(v-for="(item, i) in items" :key="item.name")
         .row
           .col.d-flex.flex-wrap.align-content-center {{ item.name }}